Fix error when creating a Response with e.g. "404"
If you call Response() with a status of "404" as a string 
(`Response('foo', status="404")`) without the reason as part of the
string, the setting of the message `msg` will fail because 
`status.split(None, 1)` will have only one value and can't be unpacked
to code and msg.

In wrappers.py, the status can be parsed to an int (line 753),
so status is set to (string) "404" and status_code is set to (int) 404.

RFC 2616 allows for an empty reason message to be sent. Without this fix,
you can't actually sent an empty message and you'll get an error if 
you're trying to build a Response object based on the response from 
a third-party server that could send an empty reason message.
(though probably rare, it happened to me, which is how I came across the
problem)

Signed-off-by: Timothée Boucher <timothee@gmail.com>